+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Bugfix: use the correct path for the file
 * Implemented getArtImage
+* Implemented getMultipleArtInfo